Add bash completion for tmuxt

+++ make_links.sh
@@ -3,8 +3,8 @@
 THIS_DIR="$(dirname $0)"
 
 THIS_DIR_RELATIVE="$(realpath --relative-to="$HOME" "$THIS_DIR")"
-for file in .bashrc .bash_aliases .vimrc .tmux.conf .tmux_powerline.conf .gitconfig; do
-	if [ -e "$HOME/$file" ]; then
+for file in .bashrc .bash_aliases .vimrc .tmux.conf .tmux_powerline.conf .gitconfig .bash_completion; do
+	if [ -e "$HOME/$file" -a ! -L "$HOME/$file" ]; then
 		mv "$HOME/$file" "$HOME/$file.old"
 	fi
 	ln -snf "$THIS_DIR_RELATIVE/$file" "$HOME/$file"
